requirements.txt
streamlit
sentence-transformers
faiss-cpu
PyMuPDF
llama-cpp-python
langchain